requirements.txt
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.5
google-search-results==2.4.2
python-dotenv==1.0.0
pandas
openpyxl
matplotlib
seaborn
pytest
faiss
numpy
